renamed file (preparing to migrate to class) htbkoo/hey-hkul-hours-native-ns-vue
